requirements-dev.txt
pyflakes
pep8
pylint
ipython
mypy
sphinx
sphinx-rtd-theme
django-stubs
